Limit ProQuest export to September 2022 and later
Why these changes are being introduced:

Only theses from the September 2022 degree period and beyond should
be exported to ProQuest.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-624

How this addresses that need:

This adds a `proquest_degree_period` scope that collects theses
with a `grad_date` equal to or greater than Sept 1, 2022, which
is added to the chains in the `partial_proquest_export` and
`full_proquest_export` scopes.

Side effects of this change:

As per usual, unrelated Report model tests have been updated to
accommodate new fixtures.