Upgrade to Rails 7
Why these changes are being introduced:

This is the first step in the Rails 7 upgrade process.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ENGX-156
* https://mitlibraries.atlassian.net/browse/ETD-625

How this addresses that need:

This commit includes the results of `rails app:update`. Files with
negligible changes (e.g., changing single quotes to double quotes)
were excluded.

Side effects of this change:

* Rather than use the rubygems version of mitlibraries-theme, we are
instead pinning it to the corresponding version in GitHub. This
was a recent team decision across all of our Rails repos that use
this gem.
* Db migrations were run in this commit, resulting in `precision: 6`
being dropped from some datetime columns in schema.db and
`precision: nil` added to others. Beginning in Rails 7.0.2,
`precision: 6` is an internal default, so it is redundant to
include it in the schema dump. Prior to Rails 6.1, precision was
not defined for datetimes in the schema, which is why `precision: nil`
is added to these columns for migrations from 6.0 and earlier.
In this case, `precision: nil` does not mean "use nil precision",
but rather "precision is not specified for this column"
  * See this PR for more information: https://github.com/rails/rails/pull/44358
  * This discussion also explains the changes to precision: https://github.com/rails/rails/issues/44571