Detect N+1 errors in development
Why are these changes being introduced:

* This app has a history of poor performance in production that we have
  been able to slowly adjust by investigating N+1 queries
* The bullet gem gives us a more structured way of detecting, logging,
  and addressing these types of issues during development so we can
  avoid pushing new problems into production

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-569
* https://mitlibraries.atlassian.net/browse/ETD-570
* https://mitlibraries.atlassian.net/browse/ETD-609
* https://mitlibraries.atlassian.net/browse/ETD-608

How does this address that need:

* Adds Bullet gem and associated configuration
* Using application in development and test envs, used the information
  provided to address N+1 queries and removed a few over eager loads.

Document any side effects to this change:

* Running Bullet as part of our test suite would be ideal, but
administrate is a bit complicated to fix the reported issues. We could
probably ignore all of the administrate errors, but it was not yet
clear if that was a good path so leaving it disabled in test is the
current plan.
* There are some over eager loads that track back to ActiveStorage. From
my reading of the documentation, we are doing the calls properly so it
wasn't clear yet how to proceed.