Migrate versions.object and versions.object_changes to JSON
Why these changes are being introduced:

The versions table, used by paper_trail, has two columns that use
YAML. PaperTrail prefers that these tables use JSON or JSONB.
In order to update paper_trail to the latest version, we need
either to update our config to use YAML.safe_load or migrate
the data to JSON.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-626
* https://mitlibraries.atlassian.net/browse/ETD-615

How this addresses that need:

This converts the columns to JSON as described here:
https://github.com/paper-trail-gem/paper_trail/blob/v14.0.0/README.md#6b-custom-serializer

Our migration uses `PaperTrail::Serializers::YAML.load`. This is
effectively the same as `YAML.safe_load`. `YAML.load`,
awhich is uggested in the readme, will no longer work due to a
long-running bug in Psych:
https://github.com/collectiveidea/audited/issues/631

We are using JSON instead of JSONB for two reasons:

1. SQLite does not support JSONB, so our local environments will
differ from Heroku; and
2. We are not actively leveraging JSONB's performance advantages in
the codebase, so there is likely no functional difference.

Long-term, it's worth considering whether we should use postgres
locally, as this would support better dev/prod parity and allow us
to use psql features like JSONB with more confidence. We discussed
this briefly in a team meeting on 3/7/23 and decided that such a
change could bring value, but is not necessary at this time.

Side effects of this change:

* One test has been updated to accommodate a value that change from
a date to a string.
* We are adding some `yaml_column_permitted_classes` to our ActiveRecord
config. This is required to run the db migration because of the
aforementioned bug in Psych. I've added a comment to remove it once
this commit is deployed.
* The migration is relatively inefficient. However, we tested it
against a clone of our prod db and found that it completed in 83.12s,
which we find acceptable.
* Running db migrations sometimes incorrectly parses the
`null: false` param into its own column. This bug occurs despite
PaperTrail's claiming to have fixed it, and it's inconsistent
(it occurs when Adam runs the migrations, but not when Jeremy does).
There's not much we can do about this except to keep an eye on it
in the future.

co-authored-by: Jeremy Prevost <jprevost@mit.edu>