Render hold status enums correctly
Why these changes are being introduced:

Following the YAML to JSON migration of the versions table,
hold statuses render in the UI as integers rather than
strings. This is because they were always incorrectly stored
as integers, but the PT YAML serializer apparently mapped them
to strings as part of the deserialization process.

In JSON, this mapping no longer occurs, so we need some additional
data handling to clean up the UI.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-629
* https://mitlibraries.atlassian.net/browse/ETD-630

How this addresses that need:

This addresses the two affected parts of the UI:

1. The hold history view, which renders `object_changes` changesets,
currently displays integers for hold status. A new hold helper
method maps the integers to the correct strings in the UI.
2. The hold admin UI has a 'date released' column that shows when
the hold's status was most recently set to 'released'.
The `Hold#date_released` method no longer evaluates this correctly,
since migrated data stores that status as the integer 2. That method
now expects either `released` or `2` as statuses.

Side effects of this change:

This fixes the bug superficially, but we need to decide what to do
about the underlying problem. (Spike ticket ETD-630 addresses this.)