Remove YAML permitted columns config
Why these changes are being introduced:

We had temporarily permitted certain data types for YAML unsafe
load to support the migration of PaperTrail data in the `versions`
table from YAML to JSON. That migration is complete, meaning we no
longer use YAML in ETD and this config is no longer needed.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-628

How this addresses that need:

This removes the unneeded config.

Side effects of this change:

None.