Remove ProQuest language from submission confirmation views
Why these changes are being introduced:

Outdated language regarding ProQuest occurs in the thesis submission
confirmation view and corresponding mailer view.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-632

How this addresses that need:

This removes the outdated text.

Side effects of this change:

An assertion that compares the enqueued email body with a
fixture has been removed (along with the fixture, which is
only used in that test). I acknowledge some coverage reduction
in removing this assertion, but I think it's inconsistent with
our philosophy of not testing views. I'm not opposed to testing
the presence of certian information in the email body, but I
don't think that a fixture is the optimal way to do so.