Enable fake auth for new staging app
Why these changes are being introduced:

We have a temporary staging app, `thesis-submit-staging-new`,
that we will use to test the new S3 bucket and SQS queues in the
AWS org. It doesn't make sense to set up this second staging
app with Touchstone, so it's best to use fake auth for testing
purposes.

Relevant ticket(s):

N/A.

How this addresses that need:

This allows the new staging app to use fake auth.

Side effects of this change:

We'll want to remove this feature once we're done testing. An inline
comment reiterates this. Longer term, it may make sense to log an
exception and crash the application if fake auth is attempted in prod.
(This is the situation that the `app_name_pattern_match?` method is
trying to prevent.)