Allow thesis processors to assign accession numbers to degree periods
Why these changes are being introduced:

In order to automate SIP processing in Archivematica, incoming
SIPs must have an accession number in their S3 keys. We can
facilitate this on the ETD side by letting thesis processors
create accession numbers and assign them to degree periods, which
in turn will enable a thesis to look up its accession number.

This PR makes some assumptions:
* Each degree period has one and only one accession number.
* Each accession number is in the format `YYYY_ddd`, where `YYYY`
is a four-digit year and `ddd` is a three-digit sequence number.
* A thesis will always have the same accession number (i.e., a
thesis' grad date will never change, or if it does, the accession
number will remain the same).

These assumptions are based on our current understanding of the
requirements. However, we expect that edge cases will arise and
and can make changes retroactively if this approach does not meet
the need. The exception would be looking up an historical
accession number if a thesis' accession number were to change,
since we are not recording that state on the Thesis model.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-589

How this addresses that need:

This adds ArchivematicaAccession and Degree Period models and
corresponding administrate dashboards. The degree periods table
is autopopulated on creation based on the grad date data in the
thesis table.

Processors can use the dashboards to create archivematica
accessions and assign them to degree periods, or create degree
periods to prepare for future thesis submissions.

Note that assigning accession numbers to theses/SIPs is explicitly
out of scope for this ticket. This will likely be a convenience
method on the Thesis model that looks up the accession number as
needed, which will also be added as a publication readiness check.

Side effects of this change:

* As noted above, this approach does not allows us to look up a
past accession number for a thesis. If there's any possibility that
a thesis' accession number will change over time, then we should
consider adding state to the Thesis or SIP models to facilitate this.
* It may be slightly confusing to introduce a Degree Period model
that only interacts with the ArchivematicaAccession model. This is
intended as a course correction for not having a Degree Period model
when we began developing this application. Degree periods are central to
the ETD data model and should likely be first class citizens.
Longer term, the grad date logic on the Thesis and Transfer models
should be refactored to reference the Degree Period model.
* Optimally, processors would be able to create and edit this data
in a single custom UI rather than in two administrate dashboards.
However, processors already use administrate to do other things. A
potential future refactor would be to ensure that processors can
do everything they need outside of administrate, so the dashboards
are only for developer use.

co-authored-by: Jeremy Prevost <JPrevost@users.noreply.github.com>