Ensure theses are not publishable if their degree period has no accession number
Why these changes are being introduced:

Because we want the accession number to be used in the preservation
workflow, we do not want a thesis without an accession number to
be published.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-596

How this addresses that need:

This adds an `accession_number` convenience method to the Thesis model
that can look up an accession based on a thesis' degree period. This
method is included in the publication checks.

The thesis processing form now shows whether it can find an accession
number. If no accession number exists, the form will render a hint that
prompts the user to create an accession number for the given degree
period.

Side effects of this change:

* A future state of this application is for the Thesis model to belong
to the Degree Period model. For now, there is no direct assocation,
so we are using ActiveRecord lookups based on the thesis' graduation
year and graduation month.
* Some fixtures have been added to allow tests related to publication
status to pass. (Certain thesis fixtures had no degree periods or
accession numbers, they could not be put into publication review.)
* The `new` action on the Archivematica Accessions dashboard has been
overwritten to prefill the degree period ID from params.