Update administrate gem
Why these changes are being introduced:

There is a new version of the administrate gem. Normally this
would occur in a regular dependency update commit, but this
version includes changes to the upstream views that we customize
in our application.

Relevant ticket(s):

N/A.

How this addresses that need:

This updates the gem and pulls in upstream changes to the layout
and `show` view.

Side effects of this change:

I am intentionally excluding one upstream change, which adds
a 'Destroy' button to the `show` view. I'm not sure this is useful
in our context, and it clutters up the UI.