Use Archivematica accession numbers in SIP AWS keys
Why these changes are being introduced:

Artefactual has requested that our AWS keys for SIPs should
include an Archivematica accession number. This will allow them
to automate processing of thesis SIPs.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-593

How this addresses that need:

This builds on previous work to rename the AWS key for a given
thesis' SIP using the accession number for that thesis. It also
adds a check for the presence of an accession number to the
`baggable_thesis?` method, in the unlikely (but possible) event
that someone tries to create a SIP for an unpublished thesis
via the Rails console.

Side effects of this change:

This will only affect theses preserved _after_ this change is
deployed in production. We do not have a plan to retroactively
change AWS keys, and stakeholders understand and have agreed to this.