Adjust AWS keys for Archivematica SIPs
Why these changes are being introduced:

We recently merged a proposed solution to this. However, looking
back at ETD-600, I realized I have the spec wrong. The current keys
are missing the graduation year and month.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-593

How this addresses that need:

This updates the keys to what Joe originally requested, so that
the SIP directory includes the grad month and is nested beneath
the grad year. For example: `etdsip/2023/June-2023_001`

Side effects of this change:

None.