Overhaul documentation
Why these changes are being introduced:

The ETD readme includes a lot of outdated information and is
lacking some information, including how to confirm that the app
works.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ENGX-227
* https://mitlibraries.atlassian.net/browse/ETD-631

How this addresses that need:

This makes several changes to the readme, including (but not
limited to):

* Adding a section on how to confirm functionality after
updating dependencies.
* Adding a section on how we manage audit trails.
* Adding a section on the processing workflow.
* Rewriting the introduction to describe what the application
currently does.
* Updating information that is out of date.
* Removing information that is inaccurate.
* Reordering information where it seems appropriate, including
changing header levels.

There is more work to be done here, but my intention is that this
will provide a solid foundation for future additions.

Side effects of this change:

* Added/adjusted line breaks for text that exceeded our line width.
* Removed the docker files that were mentioned in the readme. I
doubt we will ever use these, but I can add them back in if they
are useful to others.
* We still need to add diagrams. I'm hopeful that we'll be able to
do that using the Rails mermaid gem, but GitHub doesn't seem to
be ready for that just yet. Aside from the ERD, the various
workflows are good candidates for flow diagrams.