Disable transfers when in maintenance mode
Why these changes are being introduced:

In anticipation of migrating to new AWS infrastructure, we will
need to pause file uploads for a bit.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-649

How this addresses that need:

This adds a maintenance mode feature, toggled via the flipflop
gem, that disables and hides the thesis transfer form and changes
the welcome message on `transfer/new` to explain what is happening.
A custom message can be configured to provide additional details.

Side effects of this change:

* It is still technically possible to post a new transfer if one
were to bypass the form. I find it unlikely that transfer submitters
would try to hack this, but if it's a concern, we could try a
different approach.
* Currently, maintenance mode only disables thesis transfers. I chose
to give the feature a more generalized name in case we want to use
it to disable additional functionality in the future.