Add note about dyno scaling in pub workflow to readme
Why these changes are being introduced:

We need to scale up the dyno for the publication post-processing
job. This is not currently documented.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-650

How this addresses that need:

This adds a note about dyno scaling to the publishing workflow
section of the readme.

Side effects of this change:

None.