Adjust language in publication results email
Why these changes are being introduced:

The publication results email lists theses that were "sent to
preservation." This language is inaccurate; the theses are queued
for preservation at this point, and may not be sent to the
Archivematica bucket if the preservation job fails.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-645

How this addresses that need:

This changes the language to "Total theses queued for preservation."

Side effects of this change:

This does nothing to confirm or report on which theses were actually
sent to preservation. (ETD-646 will evaluate the best way to do this.)