Add report emails to preservation workflow
Why these changes are being introduced:

We found a bug in the preservation workflow last month that we
might have identified sooner if the preservation submission job
triggered a report email.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-651

How this addresses that need:

* Removes the Preservation Submission Prep Job
* Updates the Preservation Submission Job to accept an array of theses
* Adds a preservation results email to the Report Mailer that is
enqueued on completion of the Preservation Submission Job
* Updates the preservation rake task to convert the input thesis into
an array, so it will continue to work

Side effects of this change:

We not sure why we implemented the 'prep' job pattern: passing in
an array of theses so the actual job takes a single thesis as a
parameter. There is some risk in removing this job, since we don't
remember why it exists in the first place.