Allow manually republishing thesis in `Pending publication` status
We had a small batch of theses not end up in the SQS queue as expected.

This small change should allow us to republish each one (one at a time) to see if we
can better track down what happened (i.e. if one errors, we can assume something about it is causing this problem. If they all succeed, it was something external to an individual thesis).

This does not solve the root problem, it just allows us to more efficiently retry publication on individual theses when this app thinks they are in a publishing queue but they are not.