Unskip publication results job tests and make message IDs unique
All failing tests in the DspacePublicationResultsJob test are
related to the stubbed message that involves a bad checksum. On
further investigation, this message has the same `message_id` as
another message.

[The SQS API docs](https://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_Message.html)
specify that the `MessageId` property must be unique. It's possible
that this was always the case, but it was only enforced in more
recent versions of `aws-sdk-sqs`, which would explain why these
tests only started erroring when we tried to upgrade the gem.