Update note about confirmation emails in publishing workflow docs
Why these changes are being introduced:

The readme indicates that the last email to look for while running
the publication workflow is the preservation email. The past few
times I've run the job, it's been the MARC export email.

Relevant ticket(s):

N/A

How this addresses that need:

This generalizes the aforementioned note to check for the presence
of all three emails, before suggesting that the MARC email may
be the last one in the series.

Side effects of this change:

None.