Remove ProQuest submission field from thesis forms
Why these changes are being introduced:

The Libraries are suspending the pilot with ProQuest, and the
thesis forms should be updated.

Relevant ticket(s):

* [ETD-658](https://mitlibraries.atlassian.net/browse/ETD-658)

How this addresses that need:

This removes the ProQuest fieldset from the new and edit forms.
It also removes the author_fields partial, which includes only this
fieldset.

Note that this does not remove any additional ProQuest export
functionality, as specifically requested by stakeholders.

Side effects of this change:

Because no functionality outside of the thesis form has been
changed, thesis processors can still run ProQuest exports. There
is no apparent risk to this, and we can fully remove the feature
if stakeholders decide to do so.