Remove ProQuest consent field from receipt email
Why these changes are being introduced:

We removed the ProQuest consent field from the thesis submission
form, but forgot to remove the field from the mailer. Thus, when
a grad student submits their thesis, they see this nonexistent
field in their confirmation email.

Relevant ticket(s):

N/A

How this addresses that need:

This removes the field from the mailer.

Side effects of this change:

None.