Rails 7.0.8.7
Concurrent-ruby stopped requiring a dependency that is breaking in Rails 7.0 applications.
- https://github.com/ruby-concurrency/concurrent-ruby/pull/1062

Rails has fixed this in 7.0 branch but as it is a bugfix and not a security issue they don't plan to release a point release at this time.
- https://github.com/rails/rails/pull/54264

When we update to Rails 7.1, we can remove this pinned version of concurrent-ruby.