Wire preservation workflow to Archival Packaging Tool (APT)
Why these changes are being introduced:

DataEng has developed [APT](https://github.com/MITLibraries/archival-packaging-tool/)
as middleware between ETD and Archivematica. This new application
handles the BagIt logic, including creating bags in an S3 bucket
connected to Archivematica. Thus, much of the SIP logic in ETD is no
longer required.

Relevant ticket(s):

* [ETD-669](https://mitlibraries.atlassian.net/browse/ETD-669)

How this addresses that need:

This adds an Archivematica Payload model that effectively replaces
the SIP model. The new model constructs the payload JSON expected
by APT. Instantations of the model generate and persist this JSON
on create, along with the metadata CSV as an ActiveStorage
attachment.

The other significant change is in the Preservation Submission Job.
Previously, this job invoked the Submission Information Package
Zipper model to stream a serialized bag to S3. Now, it's
responsible for POSTing the JSON data to APT and handling the
response.

Side effects of this change:

* The tests that call APT use webmock and stubbed responses. We
would normally use VCR for external API calls, but in this case
it doesn't seem prudent to pollute the APT S3 bucket, as it's
possible the current test bucket will become the bucket we use.
* The SIP model is retained for historical purposes. This is not
ideal in terms of maintainability, but it feels important to
retain that data, at least for the time being.