Remove testing line from fake auth config
Why these changes are being introduced:

While debugging a build failure, I noticed a todo to remove a
bit of code that we used to test the new AWS org. That migration
has finished, so we no longer need it.

Relevant ticket(s):

N/A

How this addresses that need:

This removes the line of code in question.

Side effects of this change:

None.