Update publishing workflow instructions
Why these changes are being introduced:

We previously had to scale up Heroku dynos prior to running the
publishing job. Now that APT handles the bagging of SIPS, this is no
longer necessary. We have run two publishing jobs and have not yet
seen a memory spike.

Relevant ticket(s):

* [ETD-669](https://mitlibraries.atlassian.net/browse/ETD-669)

How this addresses that need:

This removes the dyno scaling instructions from the readme.

Side effects of this change:

We've only done two publishing runs since integrating APT. It's possible
that dyno scaling is still needed, and we just haven't encountered a
problem thesis yet. This is highly unlikely, however, since creating
bags was the most memory-intensive part of the publication process.